Finish up the backport of r8604 to 2.3 branch that Zarel started. Patch by stiv, closes ticket:1152
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/branches/2.3@8767 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/bitimage.c
@@ -26,7 +26,7 @@
 #include "tex.h"
 
 
-static unsigned short LoadTextureFile(const char *FileName)
+static GLuint LoadTextureFile(const char *FileName)
 {
 	iV_Image *pSprite;
 	unsigned int i;
